+++ lib/src/widgets/index.dart
@@ -1,2 +1,4 @@
 export 'bottom_navigation_builder.dart';
+export 'drawer_navigation_builder.dart';
 export 'navigator_builder.dart';
+export 'tabs_navigation_builder.dart';